// src/app/enums/users-mode.enum.ts
export enum UsersModeEnum {
  FACTORY = 'factory',
  USERS = 'users',
  SERVICES = 'services',
  AGENCY = 'agency',
  STORE = 'store',
}

export enum UserNewRoleEnum {
  CLIENT = 'Дилер',
  INSTALLER = 'Монтажник',
  PVC = 'ПВХ',
  ALUM = 'ALUM',
  TREE = 'Дерево',
  MOSQUITO = 'М/C',
  GLASS_FIBER = 'Стеклокомпозит',
  FRAMELESS_GLAZING = 'Безрамное остекление',
  BLINDS = 'Жалюзи',
  ROLLER_SHUTTERS = 'Рольставни',
  DOUBLE_GLAZED_WINDOW = 'Стеклопакеты',
  DELIVERY = 'Доставка',
  PAINTING = 'Покраска',
  CLIMBING = 'Альпинизм',
  MOVERS = 'Грузчики',
  USED_FRAMES = 'Покупка Б/У рам',
  WHOLESALE_STORE = 'Оптовый магазин',
  CLEANING = 'Клининг',
  WELDING_WORKS = 'Сварочные работы',
  GARBAGE_REMOVAL = 'Вывоз мусора',
  ADV = 'Реклама и маркетинг',
  WINDOW_FITTINGS = 'Оконная комплектация'
}